requirements.txt
PySocks==1.5.3
oauthlib==0.7.2
requests==2.6.0
requests-oauthlib==0.4.2
twitter==1.16.0
